Make tests pass when window decorations are accessible
When client side decorations are used, the frame buttons also become
accessible. This means searching just any button may return, for
example, the Minimize button. This will not close the application,
causing the test to fail.

To avoid this problem, search for a button labeled 'Close' instead of
just any button.